fix: code review items #3–#9 (Decimal refund, logging, AI assist, hygiene)
#3 — owner_views/finances.py: use Decimal instead of float for refund
     amount → cents conversion (avoids 19.99→1998 misround)

#4 — clients/services.py: replace 4 bare `except Exception: pass` blocks
     with logger.debug() so failures are traceable in logs

#5/#6 — owner_views/ai_assist.py: log exceptions server-side, return
     generic error message to client (no raw exception text exposed);
     add @require_POST to owner_naming_ai_assist and
     owner_notification_ai_assist for consistency

#7 — remove .env.dev.bak from git index; add to .gitignore

#8 — move 18 playwright test_*.py scripts from repo root →
     scripts/playwright/ (gitignored); 2 PNGs already gitignored via *.png

#9 — ai_assist.py: bail cleanly with 503 when OLLAMA_BASE_URL unset
     (previously fell back to hardcoded 192.168.1.70 IP, causing 120s
     timeout on prod)

218 tests passing.

+++ src/atletasworld/owner_views/finances.py
@@ -1,3 +1,4 @@
+from decimal import Decimal, InvalidOperation
 from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth.decorators import login_required, user_passes_test
 from django.contrib import messages
@@ -264,7 +265,7 @@ def owner_issue_refund(request, payment_id):
     try:
         kwargs = {'payment_intent': payment.stripe_payment_intent_id}
         if amount_str:
-            kwargs['amount'] = int(float(amount_str) * 100)
+            kwargs['amount'] = int(Decimal(amount_str) * 100)
         stripe.Refund.create(**kwargs)
         messages.success(request, f'Refund initiated for {payment.client} — ${payment.amount}. Status updates via webhook.')
     except stripe.error.StripeError as e:
